Drop trap instead of Use trap
If a trap is dropped on a square with a kiwi or a predator, the trap is
dropped to the ground and the kiwi or predator is killed.

Remove the use option of trap and moved to drop.

Refactored the method name trapPredator to activate trap.

Updated tests to handle dropping traps rather than using

Created method to get kiwi from location (Similar to getPredator)