Unimplemented square visibility
Square visibility has now been unimplemented, along with redundant unit
tests. This is because the new game design desires that the player can
see the map from the start. Stamina tests have also been updated to
adjust for the new stamina recharge mechanic and removale of the lose
condition for not enough player stamina