Added Terrain Images & Image logic
Heavily modified the Resource Manager, removed sound logic (for now) and
Implemented image/texture logic. Created a StretchImage class which will
stretch to fit the panel it is attached to. Added the ImageName enum,
which provides easy refrefence to images. Modified the GridSquarePanel
to display terrains with a StretchImage instance. Added textures we are
licensed to use, modify and redistrubte. Created a Sand texture for lack
of a better one. Modified the grass textures to provide a softer look