Output window and minor map change
Added a text area called outputWindow which is set by getting the output
messages from an arraylist in game called outputMessages.

Set the outputmessage to greet the user and then set that text to the
outputWindow in the contructor.

output messages have been added where an event can occur - this will
help user's understand what has happened.

in game. java a getter for the outputmessages returns the array and the
clearMessages just wipes that array.

enoughStaminaToMove is a simple variable used in the Game class that is
used to determine whether the user has enough stamina to move - this
boolean is just there for the output message.