Changed AIController design away from threading
The AIPredatorController is no longer updated by itself as a thread but
rather by a main game loop. This design change was to provide a simplier
easier to follow structure and eliminate other possible race conditions
without creating functions thread safe. Removed the abstract controller
class as the only thing that will extend it was the AIPredatorController