New textures added, screw driver removed from map
Added new textures to resource manager, image enum name, and made only a
couple be rendered from the occupantpanel because we have more than two
types of predator that we want to display.

Removed screw driver from map as nothing can break your trap in the map.
Future update could re-include the screw driver, and have it that there
is a chance a predator will break your trap when you kill them...