Pause AIController
Added a simple paused for the AI controller. The AIControllers will be
active by default, but they can be paused by running the toggle
toggleAIPaused() method.  This has been implemented to help with testing
(So predators dont run around while we try to catch them).

Trap all predators will return false if the player's move is not okay,
and also added an updategamestate at the end of this to ensure
everything is updated.